Do things correctly
