+++ AccessMaps/accessmap/dockerfiles/unweaver/Dockerfile
@@ -14,16 +14,16 @@ RUN mkdir -p /install
 WORKDIR /install
 
 # TODO: don't use poetry? It's a bit hacky
-#RUN git clone https://github.com/nbolten/unweaver.git unweaver
-
-ADD ./unweaver /install/unweaver
+RUN git clone https://github.com/nbolten/unweaver.git unweaver
 
 WORKDIR /install/unweaver
 
-RUN poetry install
+RUN git checkout $GIT_HASH \
+ && poetry install
 
 RUN poetry add osm-humanized-opening-hours \
   && poetry add pytz
 
 
 CMD ["poetry", "run", "unweaver"]
+
