--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/.pre-commit-config.yaml
@@ -1,6 +1,0 @@
-repos:
-  - repo: https://github.com/ambv/black
-    rev: stable
-    hooks:
-      - id: black
-        language_version: python3.6
