--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/.python-version
@@ -1,1 +1,0 @@
-3.6.7
