--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/example/.gitignore
@@ -1,2 +1,0 @@
-*.db
-layers/*
