--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/example/directions-wheelchair.py
@@ -1,2 +1,0 @@
-def directions(cost, nodes, edges):
-    return {"path": nodes}
