--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/example/profile-wheelchair.json
@@ -1,19 +1,0 @@
-{
-  "name": "wheelchair",
-  "args": [
-    {
-      "name": "avoidCurbs",
-      "type": "fields.Boolean()"
-    },
-    {
-      "name": "uphill",
-      "type": "fields.Number(validate=validate.Range(0, 10))"
-    },
-    {
-      "name": "downhill",
-      "type": "fields.Number(validate=validate.Range(0, 15))"
-    }
-  ],
-  "cost_function": "cost-wheelchair.py",
-  "directions": "directions-wheelchair.py"
-}
