--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/pyproject.toml
@@ -1,25 +1,0 @@
-[tool.poetry]
-name = "unweaver"
-version = "0.1.0"
-description = "A flexible, easy to run routing engine."
-authors = ["Nick Bolten <nbolten@gmail.com>"]
-license = "MIT"
-
-[tool.poetry.scripts]
-unweaver = "unweaver:cli.unweaver"
-
-[tool.poetry.dependencies]
-python = "^3.6"
-flask = "^1.0"
-webargs = "^5.3.1"
-click = "^7.0"
-entwiner = {git = "https://github.com/nbolten/entwiner.git", rev = "15296f2c0facf9f1491d025862a1b2f71cba990b"}
-shapely = "^1.6"
-
-[tool.poetry.dev-dependencies]
-black = "18.9b0"
-pre-commit = "^1.14"
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
