--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/tests/test_unweaver.py
@@ -1,5 +1,0 @@
-from unweaver import __version__
-
-
-def test_version():
-    assert __version__ == "0.1.0"
