--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/__init__.py
@@ -1,3 +1,0 @@
-__version__ = "0.1.0"
-
-from . import cli, fields, profile, server
