--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/__main__.py
@@ -1,5 +1,0 @@
-"""unweaver CLI helper entry point."""
-from unweaver import cli
-
-if __name__ == "__main__":
-    cli.unweaver()
