--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/constants.py
@@ -1,4 +1,0 @@
-"""Constants for use in submodules."""
-
-# The rectangular distance (r-tree distance) within to search for nearby edges.
-DWITHIN = 5e-4
