--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/exceptions.py
@@ -1,5 +1,0 @@
-"""Reusable exceptions."""
-
-
-class InvalidWaypoint(Exception):
-    pass
