--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/fields/__init__.py
@@ -1,1 +1,0 @@
-from .eval import Eval
