--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/fields/eval.py
@@ -1,9 +1,0 @@
-from marshmallow import fields, validate
-
-
-class Eval(fields.Str):
-    def _serialize(self, value, attr, obj):
-        pass
-
-    def _deserialize(self, value, attr, data):
-        return eval(value)
