--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/geo/__init__.py
@@ -1,1 +1,0 @@
-from .cut import cut
