--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/network_queries/__init__.py
@@ -1,1 +1,0 @@
-from .dwithin import edges_dwithin, candidates_dwithin
