--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/parsers/__init__.py
@@ -1,1 +1,0 @@
-from .parse_profile import parse_profiles
