--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/server/__init__.py
@@ -1,3 +1,0 @@
-from .app import create_app
-from .run import run_app
-from .views import add_views
