--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/server/app.py
@@ -1,18 +1,0 @@
-from flask import Flask, jsonify
-
-
-def create_app():
-    app = Flask(__name__)
-
-    # Return validation errors as JSON
-    @app.errorhandler(422)
-    @app.errorhandler(400)
-    def handle_error(err):
-        headers = err.data.get("headers", None)
-        messages = err.data.get("messages", ["Invalid request."])
-        if headers:
-            return jsonify({"errors": messages}), err.code, headers
-        else:
-            return jsonify({"errors": messages}), err.code
-
-    return app
