--- AccessMaps/accessmap/dockerfiles/unweaver/unweaver/unweaver/server/views/directions.py
@@ -1,124 +1,0 @@
-from flask import g, jsonify
-
-from ...algorithms.shortest_path import route_legs, waypoint_legs, NoPathError
-
-import click
-
-
-def blacklist_waypoints(bwp, blacklist):
-    for bbwp in bwp.values():
-        node1 = bbwp['edge']['_u']
-        node2 = bbwp['edge']['_v']
-
-        blacklist.add((node1, node2))
-        blacklist.add((node2, node1))
-
-
-def add_blacklist(blacklist, coord_list, cost_function):
-    for i in range(len(coord_list) // 4):
-        blon1 = coord_list[4 * i]
-        blat1 = coord_list[4 * i + 1]
-        blon2 = coord_list[4 * i + 2]
-        blat2 = coord_list[4 * i + 3]
-        '''
-        blacklist_legs = waypoint_legs(g.G, [[blon1, blat1], [blon2, blat2]], cost_function)
-        for i, (bwp1, bwp2) in enumerate(blacklist_legs):
-            if bwp1 is None or bwp2 is None:
-                return jsonify(
-                    {
-                        "status": "InvalidWaypointBlacklist",
-                        "msg": "Cannot route from blacklist waypoint {}".format(i + 1),
-                        "status_data": {"index": i},
-                    }
-                )
-
-            #blacklist_waypoints(bwp1, blacklist)
-            #blacklist_waypoints(bwp2, blacklist)
-        '''
-        p1 = str(precise_round(blon1, 7)) + ', ' + str(precise_round(blat1, 7))
-        p2 = str(precise_round(blon2, 7)) + ', ' + str(precise_round(blat2, 7))
-        blacklist.add((p1, p2))
-        blacklist.add((p2, p1))
-
-    return None
-
-
-def directions_view(view_args, cost_function, directions_function):
-    click.echo("Directions view endpoint hit " + str(view_args))
-
-    lon1 = view_args["lon1"]
-    lat1 = view_args["lat1"]
-    lon2 = view_args["lon2"]
-    lat2 = view_args["lat2"]
-    blacklist = set()
-    #view_args['blacklist'] = [-122.3343196, 47.599956, -122.3343161, 47.5983948]
-    if 'blacklist' in view_args:
-        err = add_blacklist(blacklist, view_args['blacklist'], cost_function)
-        if err is not None:
-            return err
-
-    #print('blacklists:', blacklist)
-
-    legs = waypoint_legs(g.G, [[lon1, lat1], [lon2, lat2]], cost_function)
-    for i, (wp1, wp2) in enumerate(legs):
-        if wp1 is None:
-            return jsonify(
-                {
-                    "status": "InvalidWaypoint",
-                    "msg": "Cannot route from waypoint {}".format(i + 1),
-                    "status_data": {"index": i},
-                }
-            )
-        if wp2 is None:
-            return jsonify(
-                {
-                    "status": "InvalidWaypoint",
-                    "msg": "Cannot route to waypoint {}".format(i + 2),
-                    "status_data": {"index": i + 1},
-                }
-            )
-
-    #print('legs', legs)
-    def cost_function_blacklist(u, v, ddict):
-        lon1, lat1 = precise_round(u.split(',')[0], 7), precise_round(u.split(',')[1], 7)
-        lon2, lat2 = precise_round(v.split(',')[0], 7), precise_round(v.split(',')[1], 7)
-
-        u = str(lon1) + ', ' + str(lat1)
-        v = str(lon2) + ', ' + str(lat2)
-        if (u, v) in blacklist:
-            #print('\tblacklisted', u, v)
-            return float('inf')
-        if (v, u) in blacklist:
-            #print('\tblacklisted', u, v)
-            return float('inf')
-
-        return cost_function(u, v, ddict)
-
-    try:
-        cost, path, edges = route_legs(g.G, legs, cost_function_blacklist)
-        print('cost of path:', cost)
-    except NoPathError:
-        return jsonify({"status": "NoPath", "msg": "No path found."})
-
-    origin = {
-        "type": "Feature",
-        "geometry": {"type": "Point", "coordinates": [lon1, lat1]},
-        "properties": {},
-    }
-    destination = {
-        "type": "Feature",
-        "geometry": {"type": "Point", "coordinates": [lon2, lat2]},
-        "properties": {},
-    }
-
-    directions = directions_function(origin, destination, cost, path, edges)
-
-    return jsonify(directions)
-
-def precise_round(num, dec):
-    num = float(num)
-    num_sign = 1
-    if num < 0:
-        num_sign = -1
-
-    return round(num * (10 ** dec) + num_sign * 0.0001) / (10 ** dec)
